Auth refactoring
