--- middleware.ts
@@ -1,11 +1,0 @@
-import NextAuth from "next-auth";
-import { authConfig } from "./auth.config";
-
-export default NextAuth(authConfig).auth;
-
-console.log("middleware loaded");
-
-// export const config = {
-//   // https://nextjs.org/docs/app/building-your-application/routing/middleware#matcher
-//   matcher: ["/((?!api|_next/static|_next/image|favicon.ico).*)"],
-// };
