+++ src/components/login/logout-button.tsx
@@ -1,5 +1,5 @@
-import { signOut } from "../../../auth";
-
+import { signOut } from "@/lib/auth/auth";
+import Button from "../ui/button";
 export default function LogoutButton() {
   return (
     <form
@@ -8,9 +8,9 @@ export default function LogoutButton() {
         await signOut();
       }}
     >
-      <button className="flex h-[48px] grow items-center justify-center gap-2 rounded-md bg-gray-50 p-3 text-sm font-medium hover:bg-sky-100 hover:text-blue-600 md:flex-none md:justify-start md:p-2 md:px-3">
-        Sign Out
-      </button>
+      <Button noBackground className="text-xs">
+        Log Out
+      </Button>
     </form>
   );
 }
