+++ src/lib/auth/auth.ts
@@ -8,6 +8,7 @@ import { getUser } from "@/lib/users";
 
 export const { auth, signIn, signOut } = NextAuth({
   ...authConfig,
+  secret: process.env.NEXTAUTH_SECRET,
   providers: [
     Credentials({
       async authorize(credentials) {
